requirements.txt
Click==7.0
Flask==1.1.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Pillow==6.1.0
redis==3.2.1
rq==1.1.0
SQLAlchemy==1.3.5
Werkzeug==0.15.4
WTForms==2.2.1
